refactor: reorganize project structure by moving core modules to a dedicated directory
