+++ src/sniffler/core/search.py
