+++ src/sniffler/core/stats.py
